Move al ini configurations to a new module called configuration
