+++ src/extension/configuration.h
@@ -1,0 +1,11 @@
+// Unless explicitly stated otherwise all files in this repository are
+// dual-licensed under the Apache-2.0 License or BSD-3-Clause License.
+//
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2021 Datadog, Inc.
+#ifndef DD_CONFIGURATION_H
+#define DD_CONFIGURATION_H
+
+void dd_configuration_startup(void);
+
+#endif // DD_CONFIGURATION_H
